// types.ts
export interface MarkupTableKey {
  label: string;
  isCitation?: boolean;
  isLink?: boolean;
}

export interface Weapon {
  Path: string;
  Kind: string;
  Name: string;
  Hands: string;
  Quality: string;
  Power: string;
  "Poise Strength": string;
  Speed: string;
  "Crit Chance": string;
  "Equip Time (s)": string;
  Description: string;
}